Updated changes with proper code standards
